split generator into multiple files
* add string and char public module under generator
* complete product up to 6 parameters
* add tuple2-6 as simplified product2-6
